vehicle-model-classifier/requirements.txt
torch==2.7.1
torchvision==0.22.1
torchaudio==2.7.1
timm==1.0.17
gradio==3.50.2
pandas==2.3.1
Pillow==10.4.0
matplotlib==3.10.3
requests==2.32.4